Move language vars and methods into new LanguageManager class
